Include the bootstrap script in the dist.  Change the package name to libdvdread (was dvdread).  Use stat64 so that using an image file of a DVD should work.  Change some more types in the interface to use int rather than int32_t, also some internal changes of types.  Make it crack the keys for all titles when opening the first as to not pause later during playback.  Make sure to use the raw device name for access under Solaris as using the CSS calls does not work on the block device.  Make sure to get an absolute path for the directory name used to search the mount table with.  Update/correct some error messages. This is what was released as libdvdread-0.9.0.

+++ libdvdread/src/play_title.c
@@ -218,7 +218,7 @@ int main( int argc, char **argv )
             /**
              * Read NAV packet.
              */
-            len = DVDReadBlocks( title, cur_pack, 1, data );
+            len = DVDReadBlocks( title, (int) cur_pack, 1, data );
             if( len != 1 ) {
                 fprintf( stderr, "Read failed for block %d\n", cur_pack );
                 ifoClose( vts_file );
@@ -269,7 +269,7 @@ int main( int argc, char **argv )
             /**
              * Read in and output cursize packs.
              */
-            len = DVDReadBlocks( title, cur_pack, cur_output_size, data );
+            len = DVDReadBlocks( title, (int)cur_pack, cur_output_size, data );
             if( len != (int) cur_output_size ) {
                 fprintf( stderr, "Read failed for %d blocks at %d\n",
                          cur_output_size, cur_pack );
